clearHeaderFilter fix for checkboxes
when clearing filters, even with `clearFilter(true)`, the filter would clear from the data but the checkbox in the header would stay checked. setting the checked value to `false` did not work either; it would de-register the input and force me to check, uncheck and recheck it just to use the filter again. `click()` handler works just fine.