Update to download.js module
Adding an always on right align setting for fields/cells that are numbers.  An argument could be made to allow the user to specify a per cell/field style, and if that's desirable I could be persuaded to update the option.  However it would be specific to cell types, for instance, if the cell value is a number, apply style, else if it's an object, apply other style, etc.