Edits from null to "" were saving
There is an issue with the if statement. It was allowing changes from null to "" to be saved even though the intension of ((cellValue === null || typeof cellValue === "undefined") && input.value !== "") (I assume) was to ignore them. They were being let through via input.value !== cellValue, because it is an "or" statement.