Allow select editor to drop up
Right now the select editor will only drop down, even at the bottom of the document, which is not always the desired behavior.
This change adds a new optional property "allowDropup" available in editorParams object.
If the property is set to true, the editor will drop up if it's height is bigger than available space at the bottom of document.