update table size on collapse toggle
without this, if you don't use a fixed height, a scrollbar will appear when opening the responsive columns because in RowManager there is a height set this.element.style.height = (this.table.element.clientHeight - otherHeight) + "px";

alternative option is in RowManager to use a minHeight style instead, that allows the table to grow automatically and let the browser handle it. This PR is probably more aligned with the current practices of Tabulator that tends to compute things and not let the browser determine size itself.