fix multiselect
getValue never returns an undefined value, therefore we check for an empty string instead. Since we need to join in all cases (not only for multiselect = true), if the initialValue is a string, we convert it to an array. As a safety measure, don't join if there is no initialValues (shouldn't be necessary but that never hurts)