properly deal with null and undefined
the current condition does not work as intended because of the || clause that use strict !== (null or undefined is always !== than input.value)

(cellValue === null || typeof cellValue === "undefined") && input.value !== "") || input.value !== cellValue

if cellValue is null, when entering the input then leaving it, it will mark the field as "edited" when in fact there was nothing done (there is no way to distinguish "null/undefined" from blank "" in the input field). Therefore, clicking the first time creates an "edited" callback, and next time, and "cancelled" callback.

this PR fixes this by treating null, undefined and "" as "" by default.

another enhancement would be to add a editorParam to allow "null" as a text value that should be converted to actual "null" value. This could be useful for header filters that should match actual "null" db values for example, but that's outside the scope of this PR.