Fix data for "row-added" event dispatch
"row-added" is dispatched for each row, so it shouldn't send the data for all the rows being added; just the current row. This change fixes the "rowAdd" history action so the redoer can successfully recreate the row. Without this fix, the redoer creates a blank row since the data is structured incorrectly.