perf: improve virtual renderer spped by append to dom in bulk
Improves performance by reducing reflow
via less style and layout calculations.

The number of rows to render is estimated upfront
using the vDomRowHeight.

We then use this as a chunk size and keep updating in chunks
until we have added enough rows to fit in the container + buffer.

This does mean is is possible we will now add more rows to
the dom, up to a maximum of one full chunk.
This is likely to be around 80 rows.