fix: ensure height is not null
With dataTree:false and a fixed height grid, if I dragged the scroll halfway down
all rows would vanish.

This is because `rowsToRender` would end up being NaN after initial render due
to `row.getHeight()` being undefined and the sum for `rowsHeight` resulting in NaN, which caused `avgRowHeight` to also be NaN.

With `rowsToRender` and `avgRowHeight` nothing rendered.

Sometimes heightInitialized would be true because setCellHeight had be called.
The row outerHeight is 0 still, we now check offsetHeight and vDomRowHeight as fallback.